new changes seem to work ... :)
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@519 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/MainFrame.cpp
@@ -1180,7 +1180,7 @@ void MainFrame::UpdateVarTree(  const wxTreeItemId& rootbranch,
         switch(var->Type()) {
           case type_float:
             {
-            DYNAMIC_CAST_VARIABLE(int,var,varf);
+            DYNAMIC_CAST_VARIABLE(float,var,varf);
             text = (boost::format("%1% %20t FLOAT %30t %2%")
                             % var->Name()
                             % (*varf->Pointer())).str();
